% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{mcData_independence}
\alias{mcData_independence}
\title{Dataset which is used internally in the package}
\format{
An object of class \code{list} of length 2.
}
\usage{
mcData_independence
}
\description{
Dataset which is used internally in the package
}
\details{
Results of the monte carlo simulation for qad according to the setting of independence.
}
\keyword{datasets}
