% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PlantDiversity}
\alias{PlantDiversity}
\title{Plant diversity}
\format{
A data frame with 140 rows and 2 variables:
\describe{
  \item{deglacYear}{estimated year of deglaciation}
  \item{div}{Shannon diversity of plant ecology}

}
}
\usage{
PlantDiversity
}
\description{
A dataset containing 140 plots with estimated year of deglaciation and plant Shannon diversity.
}
\details{
The data set contains the estimated year of deglaciation and plant diversity of 140 plots along the successional gradient of the forefield of the Ödenwinkelkees glacier (Austria).
The plots (1 to 135) were established within the glacier forefield, the plots (136 to 140) were established outside the glacier forefield (i.e., age greater than 170 years).
Since qad is invariant w.r.t. transformations the estimated year of these 5 plots is set to 1800.

#' @source \doi{10.5194/we-20-95-2020}
}
\keyword{datasets}
